[bug 1105512] Rip out old generic form
* rip out the old generic form html, js, css and view code
* remove the "unknown product" page which stunk and we don't need now
  that we have a picker
* rework a bunch of tests so the url includes the product otherwise they
  do a request and get the picker instead of the feedback form they're
  testing
* remove all the generic_feedback_form_dev related html, css, js,
  bundles and smoketests
