--- fjord/feedback/static/css/generic_feedback_dev.less
@@ -1,398 +1,0 @@
-@import 'lib/vendor-prefix.less';
-
-/* Mobile first */
-body {
-  background-color: #333;
-  color: #333;
-  font-size: 14px;
-  height: 100%;
-  line-height: 140%;
-  margin: 0;
-  padding: 0;
-}
-
-.main {
-  background-color: #f1f1f1;
-  height: 100%;
-  margin: 0 auto;
-  min-height: 100vh;
-  padding: 0;
-  position: relative;
-  width: 100%;
-}
-
-a {
-  text-decoration: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-}
-
-.deck {
-  box-sizing: border-box;
-  clear: all;
-  display: block;
-  height: 100%;
-  margin: 0;
-  padding: 5px 0 0 0;
-  width: 100%;
-}
-
-.card {
-  box-sizing: border-box;
-  display: block;
-  height: 100%;
-  margin: 0 auto;
-  padding: 0;
-  width: 100%;
-}
-
-section {
-  margin: 55px auto 0 auto;
-}
-
-.ask {
-  margin: 0 auto;
-  padding: 0 20px;
-  text-align: left;
-}
-
-body.html-rtl .ask {
-  text-align: right;
-}
-
-.inactive {
-  display: none;
-}
-
-header {
-  border: 0px none;
-  box-sizing: border-box;
-  display: block;
-  height: 40px;
-  left: 0;
-  line-height: 40px;
-  margin: 0 auto;
-  padding: 0;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  z-index: 1;
-
-  .header-content {
-    border-bottom: 1px solid #00202f;
-    .box-shadow(0 0 4px 0 rgba(0, 0, 0, 0.75));
-    .gradient(#1B3C59, #2C6191, #1B3C59);
-    align-items: center;
-    box-sizing: border-box;
-    display: block;
-    height: 40px;
-    margin: 0 auto;
-    padding: 0;
-    text-align: center;
-    vertical-align: middle;
-    width: 100%;
-
-    .left-side {
-      box-sizing: border-box;
-      display: inline;
-      float: left;
-      margin: 5px 0 0 5px;
-      padding: 0;
-      text-align: left;
-    }
-
-    .middle {
-      float: clear;
-      left: 50%;
-      margin: 0 auto;
-      padding: 0;
-      position: absolute;
-      width: 250px;
-
-      h1 {
-        color: #fff;
-        font-family: "Fira Sans","Fira Sans","Segoe UI",Frutiger,"Frutiger Linotype","Dejavu Sans","Helvetica Neue",Arial,sans-serif;
-        font-size: 14px;
-        font-weight: 600;
-        left: -50%;
-        margin: 0;
-        position: relative;
-        text-align: center;
-        text-overflow: hidden;
-        width: 100%;
-      }
-    }
-  }
-}
-
-body.html-rtl header .header-content .left-side {
-  float: right;
-  margin: 5px 5px 0 0;
-  text-align: right;
-}
-
-#back-button-container {
-  display: none;
-  height: 30px;
-  margin: 0;
-  padding: 0;
-  width: 30px;
-  z-index: 1;
-
-  #back-button {
-    cursor: pointer;
-    height: 30px;
-    margin: 0;
-    padding: 0;
-    stroke: #f0f0f0;
-    width: 30px;
-  }
-}
-
-body.html-rtl #back-button {
-  transform: rotate(180deg);
-}
-
-#sentiment {
-  margin: 0 auto 10vh auto;
-  text-align: center;
-  width: 300px;
-
-  .sentiment-button {
-    background: 0;
-    border: 0;
-    cursor: pointer;
-    display: inline-block;
-    margin: 0 auto;
-    text-align: center;
-    vertical-align: top;
-    width: 140px;
-
-    &:hover {
-      text-decoration: underline;
-    }
-
-    .face {
-      display: block;
-      height: 140px;
-      margin: 0 auto;
-      padding: 22px 0 0 0;
-      width: 120px;
-
-      &.happy-face {
-        fill: #A5C1AB;
-        stroke: #A5C1AB;
-      }
-
-      &.sad-face {
-        fill: #C3B1AB;
-        stroke: #C3B1AB;
-      }
-    }
-
-    .face-description {
-      font-weight: 600;
-      margin: 0 auto;
-      text-align: center;
-      width: 120px;
-    }
-  }
-}
-
-.input-box {
-  box-sizing: border-box;
-  padding: 20px;
-  width: 100%;
-}
-
-.private {
-  background-color: #d5d5df;
-  .border-radius(3px);
-  margin: 0 20px;
-  padding: 5px;
-}
-
-.error {
-  color: red;
-}
-
-.invalid {
-  border: 4px solid #f88;
-  .border-radius(4px);
-}
-
-#description {
-  box-sizing: border-box;
-  height: 16vh;
-  width: 100%;
-}
-
-#description-counter {
-  color: green;
-  font-size: 0.8em;
-  margin: 0;
-  text-align: right;
-  width: 100%;
-
-  &.error {
-    color: red;
-    font-weight: bold;
-  }
-
-  &.warning {
-    color: orange;
-    font-weight: bold;
-  }
-}
-
-body.html-rtl #description-counter {
-  text-align: left;
-}
-
-#id_url {
-  box-sizing: border-box;
-  width: 100%;
-}
-
-#id_email {
-  box-sizing: border-box;
-  width: 100%;
-}
-
-aside {
-  background-color: #d8e9f5;
-  .border-radius(10px);
-  border: 1px solid #a4c5db;
-  box-sizing: border-box;
-  margin: 20px;
-  padding: 10px;
-  text-align: center;
-}
-
-.btn.submit {
-  border: 0;
-  border-bottom: 3px solid #548322;
-  .border-radius(3px);
-  box-sizing: border-box;
-  color: #FFF;
-  display: block;
-  font-weight: bold;
-  .gradient(#81bc2e, #81bc2e, #629827);
-  line-height: normal;
-  margin: 10px auto;
-  outline: 0 none;
-  padding: 8px 0;
-  text-align: center;
-  text-decoration: none;
-  text-shadow: -1px -1px 0px rgba(0, 0, 0, 0.2);
-  width: 100%;
-
-  &:disabled {
-    opacity: 0.5;
-  }
-}
-
-.product-picker {
-  .content,
-  a,
-  li {
-    margin: 0;
-    padding: 0;
-  }
-  .no-products {
-    margin: 0 auto;
-    padding: 0;
-  }
-  h1 {
-    font-size: 3em;
-    font-weight: normal;
-    line-height: 110%;
-    margin: 1rem auto;
-    text-align: center;
-    width: 70%;
-  }
-  h3 {
-    font-size: 1.1rem;
-    font-weight: normal;
-    padding: 3.6rem 1rem 0px;
-  }
-  .content {
-    text-align: center;
-    overflow: visible;
-  }
-  li {
-    color: #333;
-    display: inline-block;
-    list-style: none;
-  }
-  .product-card {
-    display: inline-block;
-    background-color: white;
-    background-position: 50% 75%;
-    background-size: 60%;
-    background-repeat: no-repeat;
-    border-bottom: 1px solid #ddd;
-    border-radius: 6px;
-    box-shadow: 1px 1px 2px rgba(51, 51, 51, .7);
-    height: 7rem;
-    margin: 0.5rem;
-    vertical-align: top;
-    width: 17rem;
-  }
-}
-
-
-/* media-query based overrides */
-@media only screen and (min-width: 768px) {
-  body {
-    font-size: 16px;
-  }
-  .main {
-    width: 700px;
-  }
-  header {
-    .header-content {
-      width: 700px;
-      .middle {
-        width: 600px;
-        h1 {
-          font-size: 16px;
-        }
-      }
-    }
-  }
-  #sentiment {
-    width: 450px;
-
-    .sentiment-button {
-      margin: 0 15px;
-      width: 170px;
-    }
-  }
-  .face {
-    height: 160px;
-    padding: 25px 0 5px 0;
-    width: 135px;
-  }
-  .face-description {
-    width: 135px;
-  }
-  .product-picker {
-    .product-card {
-        background-position: 50% 100%;
-    }
-  }
-}
-
-@media only screen and (min-width: 1024px) {
-  .main {
-    margin: 0 auto;
-    width: 900px;
-  }
-  header .header-content {
-    margin: 0 auto;
-    width: 900px;
-  }
-}
