+++ fjord/feedback/templates/feedback/generic_feedback.html
@@ -1,28 +1,15 @@
 {% extends "feedback/base.html" %}
 
 {% block extra_headers %}
-  {% if waffle.flag('feedbackdev') %}
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  {% else %}
-    <meta name="viewport" content="height=480, width=320, initial-scale=1, maximum-scale=1, minimum-scale=1">
-  {% endif %}
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 {% endblock %}
 
 {% block site_css %}
-  {% if waffle.flag('feedbackdev') %}
-    {{ css('generic_feedback_dev') }}
-  {% else %}
-    {{ css('generic_feedback') }}
-  {% endif %}
+  {{ css('generic_feedback') }}
 {% endblock %}
 
 {% block site_js %}
-  {% if waffle.flag('feedbackdev') %}
-    {{ js('generic_feedback_dev') }}
-  {% else %}
-    <script src="{{ settings.STATIC_URL }}js/lib/brick-1.0.0.byob.min.js"></script>
-    {{ js('generic_feedback') }}
-  {% endif %}
+  {{ js('generic_feedback') }}
 {% endblock %}
 
 {% block page_title %}{{ _('Submit Your Feedback') }}{% endblock %}
@@ -30,9 +17,145 @@
 {% set extra_body_attrs = {'data-form-name': 'generic'} %}
 
 {% block body %}
-  {% if waffle.flag('feedbackdev') %}
-    {% include 'feedback/generic_feedback_form_dev.html' %}
-  {% else %}
-    {% include 'feedback/generic_feedback_form.html' %}
-  {% endif %}
+  <form id="responseform" action="" method="post">
+    <div class="main">
+      <header>
+        <div class="header-content">
+          <div class="left-side">
+            <div id="back-button-container" tabindex=0>
+              <svg id="back-button" viewBox="0 0 55 100">
+                <g>
+                  <path d="M 45 13 L 0 50 L 45 87" stroke-width="7" fill="none"/>
+                </g>
+              </svg>
+            </div>
+          </div>
+          <div class="middle">
+            <h1>&nbsp;</h1>
+          </div>
+        </div>
+      </header>
+
+      <div class="deck">
+        <div class="card" id="intro" data-title="{{ _('{product} feedback')|fe(product=product.display_name) }}">
+          <section>
+            <div class="ask">
+              {% trans product=product.display_name %}
+                How does {{product}} make you feel?
+              {% endtrans %}
+            </div>
+
+            <div id="sentiment">
+              <div id="happy-button" class="sentiment-button" tabindex=0>
+                <svg class="face happy-face" viewBox="0 0 100 100">
+                  <g>
+                    <!-- 100 by 100 -->
+                    <circle cx="50" cy="50" r="43" stroke-width="7" fill="none"/>
+                    <circle cx="33" cy="37" r="7" stroke-width="0" />
+                    <circle cx="67" cy="37" r="7" stroke-width="0" />
+                    <path d="M22 58 C 35 85, 65 85, 78 58" stroke-width="6" fill="none"/>
+                  </g>
+                </svg>
+                <div class="face-description">
+                  {% trans product=product.display_name %}
+                    {{product}} makes me happy
+                  {% endtrans %}
+                </div>
+              </div>
+
+              <div id="sad-button" class="sentiment-button" tabindex=0>
+                <svg class="face sad-face" viewBox="0 0 100 100">
+                  <g>
+                    <!-- 100 by 100 -->
+                    <circle cx="50" cy="50" r="43" stroke-width="7" fill="none"/>
+                    <circle cx="33" cy="37" r="7" stroke-width="0" />
+                    <circle cx="67" cy="37" r="7" stroke-width="0" />
+                    <path d="M24 72 C 35 55, 65 55, 76 72" stroke-width="6" fill="none"/>
+                  </g>
+                </svg>
+                <div class="face-description">
+                  {% trans product=product.display_name %}
+                    {{product}} makes me sad
+                  {% endtrans %}
+                </div>
+              </div>
+            </div>
+
+            <aside>
+              <div id="back-to-picker">
+                {% trans product=product.display_name, url=url('feedback') + '?' + request.GET.urlencode() %}
+                  Is {{product}} not the product you want to give feedback on?<br />
+                  <a href="{{url}}">Pick a different product</a>.
+                {% endtrans %}
+              </div>
+            </aside>
+          </section>
+        </div>
+
+        <div class="card inactive" id="moreinfo" data-focus="#description" data-title="{{ _('Details') }}" data-back-id="intro">
+          <section>
+            <div class="ask">
+              <label class="happy" for="description">
+                {% trans %}
+                  Please describe what you like.
+                  The content of your feedback will be public, so please be
+                  sure not to include personal information such as email
+                  address, passwords or phone number.
+                {% endtrans %}
+              </label>
+              <label class="sad" for="description">
+                {% trans %}
+                  Please describe your problem below and be as specific as
+                  you can.
+                  The content of your feedback will be public, so please be
+                  sure not to include personal information such as email
+                  address, passwords or phone number.
+                {% endtrans %}
+              </label>
+            </div>
+            <div class="input-box">
+              <div id="description-counter">10000</div>
+              <textarea data-max-length="10000" id="description" name="description"></textarea>
+            </div>
+
+            <div class="ask">
+              <label for="id_url">
+                {% trans %}
+                  If your feedback is related to a website, you can include it here:
+                {% endtrans %}
+              </label>
+            </div>
+
+            <div class="input-box">
+              <input class="url" id="id_url" name="url" placeholder="http://" type="text">
+            </div>
+
+            <div class="private">
+              <div class="ask private">
+                <input id="email-ok" name="email_ok" type="checkbox">
+                <label class="email-ok" for="email-ok">
+                  {% trans %}
+                    (Optional) Check here to let us contact you to follow
+                    up on your feedback. Your email address will be kept
+                    private.
+                  {% endtrans %}
+                </label>
+              </div>
+
+              <div id="email-details" class="input-box">
+                <input id="id_email" name="email" type="text" placeholder="username@example.com" disabled/>
+              </div>
+            </div>
+            <div class="input-box">
+              <button id="form-submit-btn" class="complete btn submit" disabled>{{ _('Submit feedback') }}</button>
+            </div>
+          </section>
+        </div>
+      </div>
+    </div>
+    {% for hidden in form.hidden_fields() %}
+      {{ hidden }}
+    {% endfor %}
+    {{ csrf() }}
+  </form>
 {% endblock %}
