+++ fjord/feedback/templates/feedback/picker.html
@@ -12,7 +12,7 @@
 {% endblock %}
 
 {% block site_css %}
-  {{ css('generic_feedback_dev') }}
+  {{ css('generic_feedback') }}
 {% endblock %}
 
 {% block body %}
