--- fjord/feedback/templates/feedback/unknownproduct.html
@@ -1,34 +1,0 @@
-{% extends "feedback/base.html" %}
-
-{% block page_title %}{{ _('Unknown product') }}{% endblock %}
-
-{% block site_js %}
-  <script src="{{ settings.STATIC_URL }}js/lib/brick-1.0.0.byob.min.js"></script>
-  {{ js('singlecard') }}
-{% endblock %}
-
-{% block body %}
-  <x-deck selected-index=0>
-    <x-card id="unknownproduct">
-      <x-appbar heading="{{ _('Unknown product') }}" subheading=""></x-appbar>
-
-      <section>
-        <div class="content">
-          <p>
-            {% trans product=product, url='https://bugzilla.mozilla.org/enter_bug.cgi?product=Input&rep_platform=all&op_sys=all' %}
-              Input doesn't know about "<em>{{ product }}</em>". If you think
-              this is an error, please let us know by <a href="{{  url }}">writing
-              up a bug</a>.
-            {% endtrans %}
-          </p>
-          <p>
-            {% trans url=url('feedback') %}
-              If you are here by mistake and want to leave feedback, you can
-              do so <a href="{{ url }}">on our feedback form</a>.
-            {% endtrans %}
-          </p>
-        </div>
-      </section>
-    </x-card>
-  </x-deck>
-{% endblock %}
