+++ smoketests/pages/generic_feedback_form.py
@@ -5,8 +5,10 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
 
 from pages.base import Page
+from pages.generic_feedback_picker import PickerPage
 from pages.generic_feedback_thanks import ThanksPage
 
 
@@ -14,29 +16,42 @@ class GenericFeedbackFormPage(Page):
     _page_title = 'Submit Your Feedback :: Firefox Input'
 
     _intro_card_locator = (By.ID, 'intro')
+    _picker_link_locator = (By.CSS_SELECTOR, '#back-to-picker a')
 
-    _happy_button_locator = (By.CSS_SELECTOR, '#buttons button.happy')
-    _sad_button_locator = (By.CSS_SELECTOR, '#buttons button.sad')
+    _back_locator = (By.ID, 'back-button')
+
+    _happy_button_locator = (By.ID, 'happy-button')
+    _sad_button_locator = (By.ID, 'sad-button')
 
     _moreinfo_card_locator = (By.ID, 'moreinfo')
     _description_locator = (By.ID, 'description')
+    _description_character_count_locator = (By.ID, 'description-counter')
     _url_locator = (By.ID, 'id_url')
-    _moreinfo_character_count_locator = (By.ID, 'description-counter')
-    _moreinfo_back_locator = (By.CSS_SELECTOR, '#moreinfo button.back')
-    _moreinfo_next_locator = (By.ID, 'description-next-btn')
-
-    _email_card_locator = (By.ID, 'email')
     _email_checkbox_locator = (By.ID, 'email-ok')
     _email_locator = (By.ID, 'id_email')
-    _email_back_locator = (By.CSS_SELECTOR, '#email button.back')
 
     _submit_locator = (By.ID, 'form-submit-btn')
 
     _support_page_locator = (By.LINK_TEXT, 'Firefox Support')
 
-    def go_to_feedback_page(self):
-        self.selenium.get(self.base_url + '/feedback/')
+    def go_to_feedback_page(self, product_name):
+        self.selenium.get(self.base_url + '/feedback/' + product_name + '/')
         self.is_the_current_page
+
+    def is_product(self, product_name):
+        # Make sure we're on the intro card
+        self.selenium.find_element(*self._intro_card_locator)
+
+        # Get the first ask and make sure it has the product name in
+        # it
+        first_ask = self.selenium.find_element(By.CSS_SELECTOR, 'div.ask:first-child').text
+        return product_name in first_ask
+
+    def go_to_picker_page(self):
+        picker_pg = PickerPage(self.testsetup)
+        self.selenium.find_element(*self._picker_link_locator).click()
+        WebDriverWait(self.selenium, 10).until(lambda s: picker_pg.is_the_current_page)
+        return picker_pg
 
     def type_feedback(self, feedback):
         self.selenium.find_element(*self._feedback_locator).send_keys(feedback)
@@ -52,8 +67,8 @@ class GenericFeedbackFormPage(Page):
         self.selenium.find_element(*self._sad_button_locator).click()
         self.wait_for(self._description_locator)
 
-    def click_moreinfo_back(self):
-        self.selenium.find_element(*self._moreinfo_back_locator).click()
+    def click_back(self):
+        self.selenium.find_element(*self._back_locator).click()
         self.wait_for(self._happy_button_locator)
 
     def set_description_execute_script(self, text):
@@ -72,7 +87,7 @@ class GenericFeedbackFormPage(Page):
         desc = self.selenium.find_element(*self._description_locator)
         desc.clear()
         desc.send_keys(text)
-        
+
     def update_description(self, text):
         desc = self.selenium.find_element(*self._description_locator)
         desc.send_keys(text)
@@ -81,14 +96,6 @@ class GenericFeedbackFormPage(Page):
         url = self.selenium.find_element(*self._url_locator)
         url.clear()
         url.send_keys(text)
-
-    def click_moreinfo_next(self):
-        self.selenium.find_element(*self._moreinfo_next_locator).click()
-        self.wait_for(self._email_checkbox_locator)
-
-    def click_email_back(self):
-        self.selenium.find_element(*self._email_back_locator).click()
-        self.wait_for(self._description_locator)
 
     def check_email_checkbox(self, checked=True):
         self.wait_for(self._email_checkbox_locator)
@@ -104,6 +111,11 @@ class GenericFeedbackFormPage(Page):
         email.clear()
         email.send_keys(text)
 
+    @property
+    def is_submit_enabled(self):
+        # return not 'disabled' in self.selenium.find_element(*self._submit_feedback_locator).get_attribute('class'        
+        return self.selenium.find_element(*self._submit_locator).is_enabled()
+
     def submit(self, expect_success=True):
         self.selenium.find_element(*self._submit_locator).click()
         if expect_success:
@@ -112,10 +124,6 @@ class GenericFeedbackFormPage(Page):
     @property
     def support_page_link_address(self):
         return self.selenium.find_element(*self._support_page_locator).get_attribute('href')
-
-    @property
-    def is_moreinfo_next_enabled(self):
-        return self.selenium.find_element(*self._moreinfo_next_locator).is_enabled()
 
     @property
     def is_url_valid(self):
@@ -127,8 +135,4 @@ class GenericFeedbackFormPage(Page):
 
     @property
     def remaining_character_count(self):
-        return int(self.selenium.find_element(*self._moreinfo_character_count_locator).text)
-
-    @property
-    def is_submit_feedback_enabled(self):
-        return not 'disabled' in self.selenium.find_element(*self._submit_feedback_locator).get_attribute('class')
+        return int(self.selenium.find_element(*self._description_character_count_locator).text)
