+++ smoketests/tests/generic/test_picker.py
